refactor: rename housekeeping/latexdiff.ts to avoid basename collision
src/housekeeping/latexdiff.ts and the unrelated src/latex/latexdiff.ts
shared an identical basename despite genuinely different responsibilities
(diff-output housekeeping/packing vs. diff generation), which made
filename search and "go to file" navigation ambiguous. Renamed to
packLatexdiffvc.ts, matching its primary export (runPackLatexdiffvc).
Logic is unchanged; only src/housekeeping/index.ts's re-export path
needed updating since every external consumer already goes through the
@housekeeping barrel.
